requirements.txt
opencv-python==4.11.0.86
mtcnn==1.0.0
matplotlib==3.10.0
tensorflow==2.18.0
deepface==0.0.93
tf-keras==2.18.0
scipy==1.15.1
python-dotenv==1.0.1